+++ src/main/java/com/example/tenthread/controller/PostController.java
@@ -4,7 +4,6 @@ import com.example.tenthread.dto.*;
 import com.example.tenthread.security.UserDetailsImpl;
 import com.example.tenthread.service.PostService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
@@ -47,14 +46,5 @@ public class PostController {
     public ResponseEntity<ApiResponseDto> deletePost(@PathVariable Long postId, @AuthenticationPrincipal UserDetailsImpl userDetails) throws IOException {
         ApiResponseDto result = postService.deletePost(postId, userDetails.getUser());
         return ResponseEntity.status(200).body(result);
-    }
-
-    @ExceptionHandler({IllegalArgumentException.class})
-    public ResponseEntity<ApiResponseDto> handleException(IllegalArgumentException ex) {
-        ApiResponseDto restApiException = new ApiResponseDto(ex.getMessage(), HttpStatus.BAD_REQUEST.value());
-        return new ResponseEntity<>(
-                restApiException,
-                HttpStatus.BAD_REQUEST
-        );
     }
 }
